Repurposing the repo
This repo/site was used as an experiment in publicly hosted
web applications.  The idea was that we could use it to help
in the ideation process with the stakeholders in new product
development.  That was over 2 years ago, and it never really
got off the ground.

I'm not sure what the site will be used for this time around
but I'm not sure if we are ready to give up on it yet.